Add user_service controller tests
