+++ company_service/src/main/java/com/zed/company_service/CompanyServiceApplication.java
@@ -1,0 +1,14 @@
+package com.zed.company_service;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+@EnableFeignClients
+@SpringBootApplication
+public class CompanyServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CompanyServiceApplication.class, args);
+	}
+}
