+++ company_service/src/main/java/com/zed/company_service/dto/CompanyDTO.java
@@ -1,0 +1,16 @@
+package com.zed.company_service.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+public class CompanyDTO {
+    private Long id;
+    private String name;
+    private BigDecimal budget;
+}
