+++ company_service/src/main/java/com/zed/company_service/dto/UserInfoDTO.java
@@ -1,0 +1,14 @@
+package com.zed.company_service.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserInfoDTO {
+    private String firstName;
+    private String lastName;
+    private String phoneNumber;
+}
