+++ company_service/src/main/java/com/zed/company_service/exception/AlreadyExistsException.java
@@ -1,0 +1,7 @@
+package com.zed.company_service.exception;
+
+public class AlreadyExistsException extends RuntimeException {
+    public AlreadyExistsException(String message) {
+        super(message);
+    }
+}
