+++ company_service/src/main/java/com/zed/company_service/exception/ErrorDTO.java
@@ -1,0 +1,15 @@
+package com.zed.company_service.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+
+@AllArgsConstructor
+@Setter
+@Getter
+public class ErrorDTO {
+    private String message;
+    private LocalDateTime timestamp;
+}
