+++ company_service/src/main/java/com/zed/company_service/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package com.zed.company_service.exception;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
