+++ company_service/src/main/java/com/zed/company_service/repository/CompanyRepository.java
@@ -1,0 +1,12 @@
+package com.zed.company_service.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import com.zed.company_service.entity.CompanyEntity;
+
+@Repository
+public interface CompanyRepository extends JpaRepository<CompanyEntity, Long> {
+
+    boolean existsByName(String name);
+
+}
