--- company_service/src/main/java/zed/company_service/CompanyServiceApplication.java
@@ -1,14 +1,0 @@
-package zed.company_service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-
-@EnableFeignClients
-@SpringBootApplication
-public class CompanyServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CompanyServiceApplication.class, args);
-	}
-}
