--- company_service/src/main/java/zed/company_service/dto/CompanyDTO.java
@@ -1,16 +1,0 @@
-package zed.company_service.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-public class CompanyDTO {
-    private Long id;
-    private String name;
-    private BigDecimal budget;
-}
