--- company_service/src/main/java/zed/company_service/dto/CompanyInfoDTO.java
@@ -1,13 +1,0 @@
-package zed.company_service.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-public class CompanyInfoDTO {
-    private Long id;
-    private String name;
-}
