--- company_service/src/main/java/zed/company_service/entity/CompanyEntity.java
@@ -1,39 +1,0 @@
-package zed.company_service.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import jakarta.persistence.Transient;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import zed.company_service.dto.UserInfoDTO;
-
-import java.math.BigDecimal;
-import java.util.List;
-
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Setter
-@Getter
-@Table(name = "companies")
-public class CompanyEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "name")
-    private String name;
-
-    @Column(name = "budget", nullable = false)
-    private BigDecimal budget;
-
-    @Transient
-    private List<UserInfoDTO> employees;
-}
