--- company_service/src/main/java/zed/company_service/exception/AlreadyExistsException.java
@@ -1,7 +1,0 @@
-package zed.company_service.exception;
-
-public class AlreadyExistsException extends RuntimeException {
-    public AlreadyExistsException(String message) {
-        super(message);
-    }
-}
