--- company_service/src/main/java/zed/company_service/exception/ErrorDTO.java
@@ -1,15 +1,0 @@
-package zed.company_service.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@AllArgsConstructor
-@Setter
-@Getter
-public class ErrorDTO {
-    private String message;
-    private LocalDateTime timestamp;
-}
