--- company_service/src/main/java/zed/company_service/exception/NotFoundException.java
@@ -1,7 +1,0 @@
-package zed.company_service.exception;
-
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
