--- company_service/src/main/java/zed/company_service/mapper/CompanyMapper.java
@@ -1,26 +1,0 @@
-package zed.company_service.mapper;
-
-import org.mapstruct.BeanMapping;
-import org.mapstruct.Mapper;
-import org.mapstruct.MappingConstants;
-import org.mapstruct.MappingTarget;
-import org.mapstruct.NullValuePropertyMappingStrategy;
-import zed.company_service.dto.CompanyDTO;
-import zed.company_service.dto.CreateCompanyDTO;
-import zed.company_service.dto.UpdateCompanyDTO;
-import zed.company_service.entity.CompanyEntity;
-
-import java.util.List;
-
-@Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
-public interface CompanyMapper {
-
-    CompanyDTO toCompanyDTO(CompanyEntity entity);
-
-    List<CompanyDTO> toCompanyDTOList(List<CompanyEntity> entities);
-
-    CompanyEntity toCompanyEntity(CreateCompanyDTO dto);
-
-    @BeanMapping(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE)
-    void updateEntityFromDTO(UpdateCompanyDTO dto, @MappingTarget CompanyEntity entity);
-}
