--- company_service/src/main/java/zed/company_service/repository/CompanyRepository.java
@@ -1,12 +1,0 @@
-package zed.company_service.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import zed.company_service.entity.CompanyEntity;
-
-@Repository
-public interface CompanyRepository extends JpaRepository<CompanyEntity, Long> {
-
-    boolean existsByName(String name);
-
-}
