+++ company_service/src/test/java/com/zed/company_service/CompanyServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.zed.company_service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CompanyServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
