--- company_service/src/test/java/zed/company_service/CompanyServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package zed.company_service;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CompanyServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
