+++ company_service/src/test/resources/test-schema.sql
@@ -3,13 +3,3 @@ CREATE TABLE IF NOT EXISTS companies (
     name VARCHAR(255) UNIQUE NOT NULL,
     budget NUMERIC(19, 2) NOT NULL
 );
-
-
-CREATE TABLE IF NOT EXISTS users (
-    id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL PRIMARY KEY,
-        first_name VARCHAR(255) NOT NULL,
-        last_name VARCHAR(255) NOT NULL,
-        phone_number VARCHAR(20) UNIQUE NOT NULL,
-        company_id BIGINT,
-        FOREIGN KEY (company_id) REFERENCES companies(id)
-);
