+++ user_service/src/test/java/com/zed/user_service/UserServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.zed.user_service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
