--- user_service/src/test/resources/application-test.properties
@@ -1,10 +1,0 @@
-spring.datasource.url=jdbc:h2:mem:testdb;DB_CLOSE_DELAY=-1;MODE=PostgreSQL
-spring.datasource.driver-class-name=org.h2.Driver
-spring.datasource.username=test
-spring.datasource.password=
-spring.sql.init.mode=always
-spring.sql.init.schema-locations=classpath:test-schema.sql
-spring.sql.init.data-locations=classpath:test-data.sql
-spring.jpa.hibernate.ddl-auto=none
-spring.h2.console.enabled=true
-spring.h2.console.path=/h2-console
