--- user_service/src/test/resources/application.properties
@@ -1,10 +1,0 @@
-spring.application.name=company-service
-internal.user-service.url=http://localhost:8080
-spring.datasource.url=jdbc:postgresql://localhost:5432/company_db
-spring.datasource.username=postgres
-spring.datasource.password=qzwxecrv
-spring.datasource.driver-class-name=org.postgresql.Driver
-spring.jpa.hibernate.ddl-auto=none
-spring.sql.init.mode=always
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.format_sql=true
