+++ user_service/src/test/resources/bootstrap.yml
@@ -1,0 +1,4 @@
+spring:
+  cloud:
+    config:
+      enabled: false
