+++ user_service/src/test/resources/test-schema.sql
@@ -1,10 +1,3 @@
-CREATE TABLE IF NOT EXISTS companies (
-    id BIGINT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-    name VARCHAR(255) UNIQUE NOT NULL,
-    budget NUMERIC(19, 2) NOT NULL
-);
-
-
 CREATE TABLE IF NOT EXISTS users (
     id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL PRIMARY KEY,
         first_name VARCHAR(255) NOT NULL,
